refactor(connectors): extract side-effect-free admin CLI program builder
